feat(RPMOBILE-796): Update event emitter
