--- ios/Download/DownloadEventEmitter.m
@@ -1,16 +1,0 @@
-//
-//  DownloadEventEmitter.m
-//  react-native-video
-//
-//  Created by Valerio CARMINE IENCO (KINETON) on 19/11/24.
-//
-
-#import <React/RCTBridgeModule.h>
-#import <React/RCTEventEmitter.h>
-#import <React/RCTBridge.h>
-
-@interface RCT_EXTERN_MODULE(DownloadEventEmitter, RCTEventEmitter)
-
-RCT_EXTERN_METHOD(supportedEvents)
-
-@end
